RM 1035. Added patch from a local branch: HERC-1180: Fixing an issue with cmsis
It seems that in some cases, the memory segment is not updated
due to how it clears it's internal state. We are making two
changes related to this. The first is to make the mmap PROT_EXEC
to allow the system to function as we want it. The second
is to use a volatile pointer for our A5 segment which will
provide us more confidence that we are always getting the latest
shared content in the memory segment.

(cherry picked from commit 33f3c79ad8ec6d61034f369e805a27d656dfdc93)

Signed-off-by: Sergei Poselenov <sposelenov@emcraft.com>